fix in requirements file, add GaudiGSL dependency
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@114687 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: de78d75363b5c63faba6b866e66050c98dfe9978

+++ Rec/Tr/TrackTools/src/VeloClusterPosition.h
@@ -118,7 +118,7 @@ private:
   mutable Gaudi::XYZVector m_trackDir;      /// a track direction in global ref. frame
   mutable Gaudi::XYZPoint m_gloPoint;       /// point on sensor global ref. frame
   double m_fracPos;                         /// fractional position
-  float m_corrFactor;                       /// correct difference between sigma RMS
+  double m_corrFactor;                      /// correct difference between sigma RMS
   std::string m_paraClass;                  /// use it to specify the data that wereused to tune the error parametrisations
     
 };
